- 14:00~15:05	js: [graphviz](#20250807140000) - 15:06~18:45	learn: [整理本地模型](#20250807150600) - 20:45~21:44	check: [零散笔记](#20250807204500)

+++ bailian/ccproxy.js
@@ -16,7 +16,8 @@ function short(str) {
 app.post('/v1/messages', async (req, res) => {
   const url = 'https://dashscope.aliyuncs.com/compatible-mode/v1/chat/completions';
   const inBody = JSON.stringify(req.body);
-  console.log('⬅️ Claude → Proxy', short(inBody));
+  //console.log('⬅️ Claude → Proxy', short(inBody));
+  console.log('⬅️ Claude → Proxy', inBody);
 
   // 修正字段
   const outData = {
